requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0
gunicorn==21.2.0
python-decouple==3.8
Pillow==10.1.0
reportlab==4.0.4
loguru==0.7.2
requests==2.31.0